requirements.txt
PySide6~=6.5.1.1
tinydb~=4.8.0
Pillow~=9.5.0